requirements.txt
selenium==4.8.0
webdriver-manager==3.8.6
beautifulsoup4==4.12.2
gspread==5.9.0
requests==2.31.0
matplotlib==3.7.1
pandas==2.0.2
python-dotenv==1.0.0
google-api-python-client==2.88.0
